requirements.txt
requests
beautifulsoup4
google-auth
google-auth-oauthlib
google-api-python-client